+++ source/expr.h
@@ -1,6 +1,7 @@
 #ifndef EXPR_H
 #define EXPR_H
 
+#include <stdint.h> // int64
 #include "symbol.h"
 #include "register.h"
 #include "symbol_table.h"
@@ -53,7 +54,7 @@ struct expr {
 
 	/* used by various leaf exprs */
 	const char* name;
-	int literal_value;
+	int64_t literal_value;
 	const char* string_literal;
 	struct symbol* symbol;
 
@@ -64,7 +65,7 @@ struct expr {
 struct expr* expr_create(expr_t kind, struct expr* left, struct expr* right);
 
 struct expr* expr_create_name(const char* n);
-struct expr* expr_create_integer_literal(int c);
+struct expr* expr_create_integer_literal(int64_t c);
 struct expr* expr_create_boolean_literal(int c);
 struct expr* expr_create_char_literal(char c);
 struct expr* expr_create_string_literal(const char* str);
